Implement ParameterNameDiscover for removing dependence from spring.
