+++ restler-spring-mvc/src/main/java/org/restler/spring/mvc/asm/AsmConstants.java
@@ -1,0 +1,10 @@
+package org.restler.spring.mvc.asm;
+
+class AsmConstants {
+    static final int ASM5 = 5 << 16;
+
+    static final int ACCESS_SYNTHETIC = 0x1000;
+    static final int ACCESS_BRIDGE = 0x0040;
+    static final int ACCESS_STATIC = 0x0008;
+
+}
